% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.r.distance.matern.R
\name{cov.r1.dis.matern}
\alias{cov.r1.dis.matern}
\title{Compute part of the correlation matrix based on a matern structure}
\usage{
cov.r1.dis.matern(x, xtest)
}
\arguments{
\item{x}{Training set. Either a matrix or a data frame. The same size is n and dimension is d.}

\item{xtest}{Hold-out set, of which the ouputs will be predicted.}
}
\value{
dis.tem returns a m*n by d matrix contains the m*n combinations of (x_hk - x_jk) for h <= m, j <=n, and k < d.
}
\description{
For every point in the test set, the function computes the distances of it to all points in the training set, i.e.,
(x_hk - x_ik) for h <= m (size of the test set), j <=n (size of the training set), and k < d.
}
