% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.r.distance.power.hybrid.R
\name{cov.r1.dis.power.hybrid}
\alias{cov.r1.dis.power.hybrid}
\title{Compute part of the correlation matrix based on a PowExp structure}
\usage{
cov.r1.dis.power.hybrid(x, xtest, p)
}
\arguments{
\item{x}{Training set. Either a matrix or a data frame. The same size is n and dimension is d.}

\item{xtest}{Hold-out set, of which the ouputs will be predicted.}

\item{p}{Smoothness parameters from 1 to 2.}
}
\value{
dis returns a m*n by d matrix contains the m*n combinations of (x_hk - x_jk)^p_k for h <= m, j <=n, and k < d.
}
\description{
For every point in the test set, the function computes the powered distances of it to all points in the training set
without multiplied by the sensitivity parameters, i.e.,
(x_hk - x_ik)^p_k for h <= m (size of the test set), j <=n (size of the training set), and k < d.
}
