% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.r.matern.R
\name{cov.r1.matern}
\alias{cov.r1.matern}
\title{Compute the correlation matrix based on a matern structure}
\usage{
cov.r1.matern(dis, lambda, n, m, cp)
}
\arguments{
\item{lambda}{sensitivity parameters at lambda scale.}

\item{n}{size of training set.}

\item{m}{size of test set.}

\item{cp}{smoothness parameters one of 0, 1, 2, >=3.}

\item{dis.tem}{a m*n by d matrix contains the m*n combinations of (x_hk - x_jk) for h <= m, j <=n, and k < d. 
Must be passed from cov.r1.dis.matern().}
}
\value{
r.matrix returns the correlation matrix between all points in the test set and all points in the training set.
}
\description{
The function computes the m by n correlation between points in the hold set and points in the training set.
}
