% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.r.power.hybrid.R
\name{cov.r1.power.hybrid}
\alias{cov.r1.power.hybrid}
\title{Compute the correlation matrix based on a PowExp structure}
\usage{
cov.r1.power.hybrid(dis, lambda, n, m)
}
\arguments{
\item{dis}{a m*n by d matrix contains the m*n combinations of (x_hk - x_jk)^p_k for h <= m, j <=n, and k < d. 
Must be passed from cov.r1.dis.power.hybrid().}

\item{lambda}{sensitivity parameters at lambda scale.}

\item{n}{size of training set.}

\item{m}{size of test set.}
}
\value{
r.matrix returns the correlation matrix between all points in the test set and all points in the training set.
}
\description{
The function computes the m by n correlation between points in the hold set and points in the training set.
}
