% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.matern.R
\name{dist.R.matern}
\alias{dist.R.matern}
\title{Compute part of the correlation matrix based on a matern structure}
\usage{
dist.R.matern(x)
}
\arguments{
\item{x}{Training set. Either a matrix or a data frame. The same size is n and dimension is d.}
}
\value{
d returns a n by n matrix with all elements 0.

odut returns the indices by which to place the n*(n-1)/2 correlation in the correlation matrix filled by rows.

dup returns a n*(n-1)/2 by d matrix contains the n*(n-1)/2 combinations of (x_ik - x_jk) for any i < j and k < d.
}
\description{
The function computes the distances of any two points in the training set, i.e.,  
(x_ik - x_jk) for any i < j and k < d.
}
