% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.power.hybrid.R
\name{dist.R.power.hybrid}
\alias{dist.R.power.hybrid}
\title{Compute part of the correlation matrix based on a PowExp structure}
\usage{
dist.R.power.hybrid(x, p)
}
\arguments{
\item{x}{Training set. Either a matrix or a data frame. The same size is n and dimension is d.}

\item{p}{Smoothness parameters from 1 to 2.}
}
\value{
d returns a n by n matrix with all elements 0.

odut returns the indices by which to place the n*(n-1)/2 correlation in the correlation matrix filled by rows.

dup returns a n*(n-1)/2 by d matrix contains the n*(n-1)/2 combinations of (x_ik - x_jk)^p_k for any i < j and k < d.
}
\description{
The function computes the powered distances of any two points in the training set without multiplied by the sensitivity parameters, i.e.,  
(x_ik - x_jk)^p_k for any i < j and k < d.
}
