% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.post.power.hybrid.R
\name{log.post.power.hybrid}
\alias{log.post.power.hybrid}
\title{Compute Four components of a GP with a PowExp structure}
\usage{
\method{log}{post.power.hybrid}(d, odut, dup, y, lambda)
}
\arguments{
\item{d}{a n by n matrix with all elements 0. Must be passed from dist.R.power.hybrid().}

\item{odut}{the indices by which to place the n*(n-1)/2 correlation 
in the correlation matrix filled by rows. Must be passed from dist.R.power.hybrid()}

\item{dup}{a n*(n-1)/2 by d matrix contains the n*(n-1)/2 combinations 
of (x_ik - x_jk)^p_k for any i < j and k < d. Must be passed from dist.R.power.hybrid()}

\item{y}{True outputs of x. Size is n.}

\item{lambda}{sensitivity parameters at lambda scale.}
}
\value{
R returns the correlation matrix.

logpost returns the posterior probability at log scale without the normalization constant.

mu returns the estimated constant mean of the GP.

sigma returns the estimated variance of the GP.
}
\description{
The function computes the posterior probability at log scale without the normalization constant, the correlation matrix of 
points in the training set, the estiamted constant mean and the variance of the GP.
}
