% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.power.hybrid.R
\name{pred1.power.hybrid}
\alias{pred1.power.hybrid}
\title{Compute the predicted mean and variance based on GP with a PowExp structure}
\usage{
pred1.power.hybrid(d, odut, dup, dis, y, lambda)
}
\arguments{
\item{d}{a n by n matrix with all elements 0. Must be passed from dist.R.power.hybrid().}

\item{odut}{the indices by which to place the n*(n-1)/2 correlation 
in the correlation matrix filled by rows. Must be passed from dist.R.power.hybrid()}

\item{dup}{a n*(n-1)/2 by d matrix contains the n*(n-1)/2 combinations 
of (x_ik - x_jk)^p_k for any i < j and k < d. Must be passed from dist.R.power.hybrid()}

\item{dis}{a m*n by d matrix contains the m*n combinations of (x_hk - x_jk)^p_k for any for h <= m, j <=n, and k < d.
Must be passed from cov.r1.dis.power.hybrid().}

\item{y}{True outputs of x. Size is n.}

\item{lambda}{sensitivity parameters at lambda scale.}
}
\value{
res returns the predicted mean.

v.term2 return the predicted variance.
}
\description{
The function returns the predicted mean and variance.
}
