% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmat.R
\name{repmat}
\alias{repmat}
\title{The function is the R equivalent of repmat (matlab)}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{is a mtrix.}

\item{m}{is the number of repeats in the direction of row.}

\item{n}{is the number of repeats in the direction of column.}
}
\value{
the resulting matrix.
}
\description{
The function is the R equivalent of repmat (matlab)
}
