% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBMCMC.matern.hybrid.R
\name{runBMCMC.matern.hybrid}
\alias{runBMCMC.matern.hybrid}
\title{Gaussian process with a hybrid matern (Matern-Hybrid) correlation structure.}
\usage{
runBMCMC.matern.hybrid(nmcmc = 10000, burn = 4000, thin = 10, x, y, xtest,
  lambda.ini, lambda.w.ini, cp)
}
\arguments{
\item{nmcmc}{Length of the MCMC chain. Default is 10,000.}

\item{burn}{Burn-in period. Default is 4,000.}

\item{thin}{Thinning parameter. Default is 10.}

\item{x}{Training set. Either a matrix or a data frame.}

\item{y}{True outputs of x.}

\item{xtest}{Hold-out set, of which the ouputs will be predicted. Must have the same dimension as x.}

\item{lambda.ini}{Initial values of the sensitivity parameters, at \eqn{\lambda} scale.}

\item{lambda.w.ini}{Initial values of the lengths for adaptive MCMC in estimating the sensitivity parameters, at \eqn{\lambda} scale.}

\item{cp}{Smoothness parameters one of 0, 1, 2, >=3. Must be a vector with the same dimension as x.}
}
\value{
pred.y return the predicted mean.

pred.var return the predicted variance.

accept.rate.lambda return a vector (1*d) of the acceptance rate of lambda.

mcmc.matrix.lambda return a matrix (nmcmc * d) of the sampled lambda.
}
\description{
Gaussian process with a hybrid matern (Matern-Hybrid) correlation structure.
}
