% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sd.R
\name{s.square1}
\alias{s.square1}
\title{Predicted variance of a GP}
\usage{
s.square1(sigma, r, R)
}
\arguments{
\item{sigma}{The estimated variance of the GP. Must be passed from log.post.power.hybrid().}

\item{r}{The m by n correlation matrix of all points in the training set and all points in the test set.
Must be passed from cov.r1.power.hybrid().}

\item{R}{The correlation matrix of the training set. Must be passed from log.post.power.hybrid().}
}
\value{
rett returns the predicted variances of all points in the test set.
}
\description{
The function computes the predicted variances.
}
